zSwitch Package
Place switch areas where you want the switch to occur. Each area should
have the zSwitch script attached, link to the zSwitch Area that you
want to move to. For rightOrLeft, put 1 to appear on the right side (X+
direction) of the switch area, any other number will put object on the
left side (X- direction). Only tested this with a 1x1x1 cube, may need
to increase the offset in script if experiencing infinite switching.